feat: get app`s bitcoin balance

+++ app/routers/admin.py
@@ -4,6 +4,8 @@ from aiogram.filters import Command
 
 from app.database.queues.get_user import get_user
 
+from app.bot.get_balance import get_balance
+
 from app.keyboards.admin import admin_keyboard
 from app.keyboards.main import main_keyboard
 
@@ -17,7 +19,14 @@ async def admin_panel(message: Message) -> None:
         user = await get_user(message.from_user.id)
 
         if user[5] is True:
-            await message.answer('Вход в панель администратора ✅', reply_markup=admin_keyboard())
+            await message.delete()
+
+            balance = await get_balance()
+
+            content = 'Вход в панель администратора 🔑\n' \
+                      f'Баланс BTC кошелька приложения: {balance} ₿'
+
+            await message.answer(content, reply_markup=admin_keyboard())
         else:
             await message.answer('Ошибка входа в панель администратора ❌')
     except Exception as e:
@@ -26,6 +35,29 @@ async def admin_panel(message: Message) -> None:
 
 @admin_router.message(F.text == 'Выйти из админ панели 🔙')
 async def exit_admin_panel(message: Message) -> None:
-    await message.delete()
-    
-    await message.answer('Вы вышли из панели администратора', reply_markup=main_keyboard())
+    try:
+        user = await get_user(message.from_user.id)
+
+        if user[5] is True:
+            await message.delete()
+
+            await message.answer('Вы вышли из панели администратора', reply_markup=main_keyboard())
+        else:
+            pass
+    except Exception as e:
+        print(f'Exit admin panel error: {e}')
+
+
+@admin_router.message(F.text == 'Пополнить BTC кошелёк ₿')
+async def deposit_btc(message: Message) -> None:
+    try:
+        user = await get_user(message.from_user.id)
+
+        if user[5] is True:
+            await message.delete()
+
+            pass
+        else:
+            pass
+    except Exception as e:
+        print(f'Deposit BTC error: {e}')
